Move sensor-related stuff to internal/.
