+++ cartographer/cloud/internal/handlers/add_imu_data_handler.cc
@@ -21,8 +21,8 @@
 #include "cartographer/cloud/internal/sensor/serialization.h"
 #include "cartographer/cloud/proto/map_builder_service.pb.h"
 #include "cartographer/common/make_unique.h"
-#include "cartographer/sensor/dispatchable.h"
 #include "cartographer/sensor/imu_data.h"
+#include "cartographer/sensor/internal/dispatchable.h"
 #include "google/protobuf/empty.pb.h"
 
 namespace cartographer {
