+++ cartographer/mapping/internal/2d/pose_graph/constraint_builder_2d.h
@@ -36,8 +36,8 @@
 #include "cartographer/mapping/pose_graph/proto/constraint_builder_options.pb.h"
 #include "cartographer/mapping/pose_graph_interface.h"
 #include "cartographer/metrics/family_factory.h"
+#include "cartographer/sensor/internal/voxel_filter.h"
 #include "cartographer/sensor/point_cloud.h"
-#include "cartographer/sensor/voxel_filter.h"
 
 namespace cartographer {
 namespace mapping {
