+++ cartographer/mapping/internal/2d/pose_graph_2d.cc
@@ -33,7 +33,7 @@
 #include "cartographer/common/math.h"
 #include "cartographer/mapping/pose_graph/proto/constraint_builder_options.pb.h"
 #include "cartographer/sensor/compressed_point_cloud.h"
-#include "cartographer/sensor/voxel_filter.h"
+#include "cartographer/sensor/internal/voxel_filter.h"
 #include "glog/logging.h"
 
 namespace cartographer {
