+++ cartographer/mapping/internal/3d/pose_graph/constraint_builder_3d.h
@@ -39,8 +39,8 @@
 #include "cartographer/mapping/trajectory_node.h"
 #include "cartographer/metrics/family_factory.h"
 #include "cartographer/sensor/compressed_point_cloud.h"
+#include "cartographer/sensor/internal/voxel_filter.h"
 #include "cartographer/sensor/point_cloud.h"
-#include "cartographer/sensor/voxel_filter.h"
 
 namespace cartographer {
 namespace mapping {
