+++ cartographer/mapping/map_builder.cc
@@ -30,10 +30,10 @@
 #include "cartographer/mapping/internal/3d/pose_graph_3d.h"
 #include "cartographer/mapping/internal/collated_trajectory_builder.h"
 #include "cartographer/mapping/internal/global_trajectory_builder.h"
-#include "cartographer/sensor/collator.h"
+#include "cartographer/sensor/internal/collator.h"
+#include "cartographer/sensor/internal/trajectory_collator.h"
+#include "cartographer/sensor/internal/voxel_filter.h"
 #include "cartographer/sensor/range_data.h"
-#include "cartographer/sensor/trajectory_collator.h"
-#include "cartographer/sensor/voxel_filter.h"
 #include "cartographer/transform/rigid_transform.h"
 #include "cartographer/transform/transform.h"
 #include "glog/logging.h"
