+++ cartographer/sensor/internal/trajectory_collator.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 The Cartographer Authors
+ * Copyright 2018 The Cartographer Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,28 +14,31 @@
  * limitations under the License.
  */
 
-#ifndef CARTOGRAPHER_SENSOR_COLLATOR_H_
-#define CARTOGRAPHER_SENSOR_COLLATOR_H_
+#ifndef CARTOGRAPHER_SENSOR_INTERNAL_TRAJECTORY_COLLATOR_H_
+#define CARTOGRAPHER_SENSOR_INTERNAL_TRAJECTORY_COLLATOR_H_
 
-#include <functional>
 #include <memory>
 #include <unordered_map>
-#include <unordered_set>
 #include <vector>
 
 #include "cartographer/sensor/collator_interface.h"
-#include "cartographer/sensor/data.h"
-#include "cartographer/sensor/ordered_multi_queue.h"
+#include "cartographer/sensor/internal/ordered_multi_queue.h"
 
 namespace cartographer {
 namespace sensor {
 
-class Collator : public CollatorInterface {
+// Waits to see at least one data item for all sensor ids and dispatches data
+// in merge-sorted order. Contrary to 'Collator', it does not wait for other
+// trajectories.
+// Also contrary to 'Collator', whose output is deterministic, the sequence in
+// which data is dispatched is not sorted, so non-deterministic input sequences
+// will result in non-deterministic output.
+class TrajectoryCollator : public CollatorInterface {
  public:
-  Collator() {}
+  TrajectoryCollator() {}
 
-  Collator(const Collator&) = delete;
-  Collator& operator=(const Collator&) = delete;
+  TrajectoryCollator(const TrajectoryCollator&) = delete;
+  TrajectoryCollator& operator=(const TrajectoryCollator&) = delete;
 
   void AddTrajectory(int trajectory_id,
                      const std::unordered_set<std::string>& expected_sensor_ids,
@@ -50,14 +53,13 @@ class Collator : public CollatorInterface {
   common::optional<int> GetBlockingTrajectoryId() const override;
 
  private:
-  // Queue keys are a pair of trajectory ID and sensor identifier.
-  OrderedMultiQueue queue_;
+  std::unordered_map<int, OrderedMultiQueue> trajectory_to_queue_;
 
   // Map of trajectory ID to all associated QueueKeys.
-  std::unordered_map<int, std::vector<QueueKey>> queue_keys_;
+  std::unordered_map<int, std::vector<QueueKey>> trajectory_to_queue_keys_;
 };
 
 }  // namespace sensor
 }  // namespace cartographer
 
-#endif  // CARTOGRAPHER_SENSOR_COLLATOR_H_
+#endif  // CARTOGRAPHER_SENSOR_INTERNAL_TRAJECTORY_COLLATOR_H_
